[FIX] Add tty parameter when creating a new docker
Logs from a container being retrieved from Docker API are coming with some garbage. This problem is being treated by eliminating the undesired output via a JSON regexp. However, when these logs are too big, the message is chunked and more garbage are inserted into the JSON.

To fix this issue, a container is now being created using the TTY parameter.

Closing issue #35