[FEAT] Add a new Regexp into analysis.go
Initial idea was to use CI_REPOSITORY_URL from the runner from GitlabCI
to use it as a parameter. However, by doing some checks, it is possible
to realize that it is better to force HuskyCI user to hardcode its
repository URL in this format git@github.com:globocom/husky.git.

Closes issue #70