[REFACT] Make code compliant with generic DB
This is huge refactoring so all function calls for data retrives can
be made through the Requests interface. As a default, a Mongo struct
is created. The name of the environment variables were changed so it can
be consistent with a generic DB and not associated with Mongo. If the
HUSKYCI_DATABASE_TYPE is configured with postgres value, a struct that
implements this database will be instantiated.