Make database calls generic (#400)
* [FEATURE] Add psql package and start psql support

* [FEATURE] Implement unit tests in sql functions

* [REFACT] Update packages

* [FEATURE] Defining new interface for DB requests

It will make the functions that retrieve some data from DB generic so
new DBs can be used in the future with the implementation of the defined
interface.

* [FEATURE] Make improvements to attend lint rules

* [REFACT] Make code compliant with generic DB

This is huge refactoring so all function calls for data retrives can
be made through the Requests interface. As a default, a Mongo struct
is created. The name of the environment variables were changed so it can
be consistent with a generic DB and not associated with Mongo. If the
HUSKYCI_DATABASE_TYPE is configured with postgres value, a struct that
implements this database will be instantiated.

* [VERSION] Changes in packages versions

* [FIX] Adjust defer func so it can fit the lint

* [FIX] Make the changes requested in the PR

All changes indicated on the PR by my comment are done in this commit.