[REFACT] Identify special cases in SELECT query
This changes allow RetrieveFromDB function to find data returned from
DB as JSON. If it is, it will convert for an interface type so it could
be recovered on the struct associated with this JSON.

The other change is the possibility to pass the column names there are
array of strings. The sql driver doesn't convert them to a slice type.
It will be converted using ConvertStringToSlice function so it could be
returned to the response struct and recovered on this.