enhancement: make security test timeout configurable via env variable
Security Test timeouts are currently hardcoded in API's
config.yml.

Provide user options to use a environment variable
HUSKYCI_CLIENT_TESTS_TIMEOUT in the client code. All requests
send to the API henceforth will contain that timeout.

On the API side, if a non zero custom timeout is detected, it
shall override the default timeouts for the security tests.

Signed-off-by: Dwaipayan Ray <dwaipayanray1@gmail.com>